Move service/common/api/http -> service/common/httpencode
