+++ internal/pkg/service/common/httpencode/encode.go
@@ -1,4 +1,4 @@
-package http
+package httpencode
 
 import (
 	"context"
